[iss-145] Squashed commit of the following:
commit 9a8cfff832673c05b893cb4e7161606c8910483d
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Thu Sep 9 09:52:41 2021 -0300

    Enable interleaved test case.

commit 88ede856db1ad9d56c89125a57995adf984fb257
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Thu Sep 9 09:52:08 2021 -0300

    Updated interleaved gt data.

commit f24fa67bdfdde78738657787c0b13746236d91d5
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Thu Sep 9 09:51:34 2021 -0300

    Handle recursive case in sb deps.

commit a8383f9e09251d30a5753a5cc85cb6c837e566db
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Thu Sep 9 09:50:20 2021 -0300

    Added const end iterator and non-zero slopes method to lmap exp.

commit b3ddff989bd00bc659069c627b3d6cdf561de661
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Thu Sep 9 09:49:11 2021 -0300

    Updated var deps definition.

commit 0f30818b47fe40d7f79cdad5b3fa19dc9bf1b084
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Thu Sep 9 09:48:35 2021 -0300

    Added unique ids for equation  graph nodes.

commit 9c446664c6afbce4d3a1865aadac496e5697b2f0
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Thu Sep 9 09:47:50 2021 -0300

    Added atomize method to sets

commit 5e84869867f6f16f57561ab55b6e3887842bc5b7
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Aug 31 16:18:51 2021 -0300

    Update test C++ std.

commit 7f92339d9986d82a2d6bb625f6e86a2ead609ec7
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Mon Aug 30 16:27:24 2021 -0300

    Updated GT files with new Jacobian structs.

commit e18900013edcfb965170007caafb9032f58340c4
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Mon Aug 30 16:22:22 2021 -0300

    Temp removed interleaved test until recursion is enabled again.

commit f3c40cd524a831bc1bb60f017ad0504e58901f17
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Mon Aug 30 16:20:33 2021 -0300

    Use interval refs to avoid segfaults.

commit add964a51665474938e19f62c9e94a495fcdb753
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Mon Aug 30 16:20:08 2021 -0300

    Added isScalar method to pwl_map_values.

commit fb7d9a86d48d2db0fd4bb49e48731be3f758135c
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Mon Aug 30 16:17:35 2021 -0300

    Removed dead code.

commit 5f84a44851f10e17799f456bc9dcab232265d74b
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Mon Aug 30 16:17:23 2021 -0300

    Fix assert for scalar expressions.

commit f279c7892193047f9a2062ea8c7bd5a981747715
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Mon Aug 30 16:17:07 2021 -0300

    Removed dead code.

commit a5927be33ea94bdbaaca2c07c592623446036467
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Mon Aug 30 16:16:55 2021 -0300

    Fixed multidim set interval builder.

commit f27400e4edc3cd1b252992fb076fcc1092a34ec8
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Mon Aug 30 16:16:08 2021 -0300

    Added atomic sets ref method.

commit 395a80126d2495bdbdaaed279d4a914bd891e1bf
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Mon Aug 30 16:15:50 2021 -0300

    Generate Index expressions for guards.

commit d6d172e7f5dd73c24e822a060550790d9fddb686
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Mon Aug 30 16:15:26 2021 -0300

    Generate Index expressions for guards.

commit e81cd3c8c5265ec7d767315208f3f1893876fb27
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:26:01 2021 -0300

    Modified sb dependencies algorithm to use sbg structs.

commit dae4a8575dce9421b679fc716d39d5e9ef47958b
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:25:32 2021 -0300

    Make Jacobian builder use sbg structs.

commit ff99a8e80c1c887d03121ad6b99de351b0de10c8
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:25:14 2021 -0300

    Make Jacobian matrix builder use sbg structs.

commit 801b5ad9a2acca6e09c2dca00eaaabadc9c0f156
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:24:42 2021 -0300

    Added sbg sd graph builder.

commit 1248796d4f9b16674c47e620364721a1c93495bc
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:24:23 2021 -0300

    Fix sd graph builder.

commit 5b59a8a4959370874793de834642d382d7f9ef16
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:23:54 2021 -0300

    Use Deps::IndexShoft definition.

commit f0b1b1c86fe6af2976b20fda85e65c76a613752c
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:21:14 2021 -0300

    Added visitor to get map values.

commit b0cdfbce6f5e15b88913a82560ef7ff68baa61f2
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:20:48 2021 -0300

    Added logs path to logger.

commit dfe29c7ac578a67f5be639dfea6d5e78de7b2ff5
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:20:29 2021 -0300

    Avoid size type warning.

commit 21b2ae1115e66eb84fb41bd8ace9b94a2fd964f8
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:20:06 2021 -0300

    Expose partial eval expression methods.

commit bfbd2e69387ec46da632fe35760ad442368dfb33
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:19:34 2021 -0300

    Use sets in occurs.

commit 2160c0b213a98de5f062ef9408be19c330e2fc75
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:18:55 2021 -0300

    Define order in mmo expressions.

commit d654d620bc53d8b9da94dbf026a75b71957de767
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:18:28 2021 -0300

    Eval ints in constant expressions.

commit 10d243f990ab3f9e39406f8a422dafa5df86a697
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:18:00 2021 -0300

    Cleanup original graph definition

commit 6ca9067880cffee03eacc422eae25eaaace98d1b
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:17:00 2021 -0300

    Build range from sbg set.

commit 10458163590a60ce4c563dabe2bbc3c0f7baec8f
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:16:24 2021 -0300

    Added max dim to symbol table.

commit 194f4079e9682e0c9cbb1f3bd331fbc3c2638c18
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:15:25 2021 -0300

    Use expression set instead of lists.

commit c53572ac22890d06130378c6c339c547175ce36d
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:14:36 2021 -0300

    Added order relation to AST nodes.

commit 02c6f2bfc13d4ceec24d3177727e81d7cc645d8f
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:13:59 2021 -0300

    Added order relation macros.

commit 3fc99a08bb4987527a179467d8077faf84e6d78c
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:13:32 2021 -0300

    Compile sbg graphs,

commit 40aa8d8659752858997db3bfe4de1e7f6f2a9e75
Author: joaquin.f.fernandez <joaquin.f.fernandez@gmail.com>
Date:   Tue Jul 13 11:13:01 2021 -0300

    Added sbg definitions.

+++ src/mmoc/util/visitors/replace_index.cpp
@@ -44,7 +44,7 @@ AST_Expression ReplaceIndex::foldTraverseElement(AST_Expression exp)
       AST_ExpressionListIterator it;
       int i = 0;
       AST_ExpressionList l = newAST_ExpressionList();
-      assert(indexes->size() == _usage.Size());
+      assert(indexes->size() == (size_t)_usage.Size());
       foreach (it, indexes) {
         if (_usage.isUsed(i)) {
           string var = _range.iterator(i, _range_idxs);
